requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cloudinary==1.29.0
Django==4.0.4
django-crispy-forms==1.14.0
Pillow==9.1.0
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
